Lua 5.1 doesn't support goto labels. (#88)
Unfortunately, LuaJIT supports Lua 5.1 with a slightly random set of Lua
5.2 extensions, which means it's possible to write code which doesn't
work on those platforms where only Lua 5.1 (and not LuaJIT) can be
used (e.g. OpenBSD).

In this case "goto ::x::" isn't a Lua 5.1 feature. Fortunately the
change in control flow in this case is easy and, in my opinion, fairly
natural.