Fixed errors being thrown when signature is nil (#116)
Somtimes the "signature" variable is null, and the plugin throws errors
that slow down when you are typing and are just generally annoying.

This happens, for instance, when you type in more arguments than the
function signature actually has. Of course, in this case the user is in
a way wrong, but the constant error messages make it difficult to fix the
mistake.

This adds some nil guards to prevent the errors form popping up.