ignore signature results for other buffers (#233)
Language servers can be slow to respond to the
`textDocument/signatureHelp` request. During that time, the user could
have moved to another buffer. The signature would be shown for the
current buffer even though it was meant for some other buffer.

Let's only show the signature if it is for the current buffer.

This bugfix is cherry-picked from the Neovim core `signature_help` LSP
handler [0].

[0]: https://github.com/neovim/neovim/pull/21121